requirements.txt
telethon>=1.30.0
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
python-dotenv>=1.0.1
python-socks[asyncio]>=2.7.2
